refactor: strengthen test infrastructure and consolidate shared logic
- Add tests/helpers/schemas.ts with common GraphQL schema constants shared
  across all five validator spec files (PrimitiveInput, ArrayInput, etc.)
- Add tests/helpers/plugin.ts with a runPlugin() wrapper to reduce
  buildSchema + plugin() boilerplate in tests
- Add expectTypeScriptToCompile checks to myzod, valibot, and yup primitive
  tests (myzod/valibot had zero compile checks; this adds a baseline)

- Extract src/zod/shared.ts: 22 module-level functions (≈250 lines) that were
  100% identical in zod/index.ts and zodv4/index.ts; both files now import from
  the shared module (similarity-ts reported 100% match on generateFieldTypeZodSchema
  and defaultValueExpression among others)
- Add src/lazy.ts with buildMaybeLazy() — the core isComplex check that drives
  lazy-reference wrapping was duplicated across all five validators; each now
  passes its own library prefix as a callback
- Add src/scalar.ts with buildScalarSchema() — the scalarSchemas → typeMap →
  defaultScalarTypeSchema → fallback pattern was ~84% similar across all five
  validators; yup's .defined() wrapping is handled via the wrapCustom option

All 357 tests pass unchanged.
