+++ src/zodv4/index.ts
@@ -1,43 +1,42 @@
 import type { Types } from '@graphql-codegen/plugin-helpers';
 import type {
-  ConstValueNode,
   EnumTypeDefinitionNode,
   FieldDefinitionNode,
   GraphQLSchema,
   InputObjectTypeDefinitionNode,
-  InputObjectTypeExtensionNode,
   InputValueDefinitionNode,
   InterfaceTypeDefinitionNode,
-  NameNode,
   ObjectTypeDefinitionNode,
-  TypeNode,
   UnionTypeDefinitionNode,
 } from 'graphql';
 
 import type { ValidationSchemaPluginConfig } from '../config.js';
 import type { Visitor } from '../visitor.js';
-import { resolveExternalModuleAndFn } from '@graphql-codegen/plugin-helpers';
-import { convertNameParts, DeclarationBlock, indent } from '@graphql-codegen/visitor-plugin-common';
+import { DeclarationBlock, indent } from '@graphql-codegen/visitor-plugin-common';
+import { Kind } from 'graphql';
 import {
-  isEnumType,
-  isInputObjectType,
-  isScalarType,
-  Kind,
-  valueFromASTUntyped,
-} from 'graphql';
-import { buildApi, formatDirectiveConfig } from '../directive.js';
-import {
-  escapeGraphQLCharacters,
   InterfaceTypeDefinitionBuilder,
-  isListType,
-  isNamedType,
-  isNonNullType,
   ObjectTypeDefinitionBuilder,
 } from '../graphql.js';
 import { BaseSchemaVisitor } from '../schema_visitor.js';
 import { buildZodOperationSchemas } from '../zod/operation.js';
-
-const anySchema = `definedNonNullAnySchema`;
+import {
+  anySchema,
+  applyDirectives,
+  buildObjectExpression,
+  buildObjectReturn,
+  enumDefaultTypeName,
+  enumDefaultValueName,
+  generateFieldZodSchema,
+  generateFieldTypeZodSchema,
+  isOneOfInputObject,
+  maybeLazy,
+  schemaDepthParameter,
+  schemaDepthVariable,
+  unionLiterals,
+  withDescription,
+  zod4Scalar,
+} from '../zod/shared.js';
 
 export class ZodV4SchemaVisitor extends BaseSchemaVisitor {
   constructor(schema: GraphQLSchema, config: ValidationSchemaPluginConfig) {
@@ -326,276 +325,6 @@ export class ZodV4SchemaVisitor extends BaseSchemaVisitor {
   }
 }
 
-function generateFieldZodSchema(config: ValidationSchemaPluginConfig, visitor: Visitor, field: InputValueDefinitionNode | FieldDefinitionNode, indentCount: number, depthVariable?: string): string {
-  const gen = generateFieldTypeZodSchema(config, visitor, field, field.type, undefined, true, false, depthVariable);
-  return indent(`${field.name.value}: ${withDescription(config, field, maybeLazy(visitor, field.type, gen))}`, indentCount);
-}
-
-function generateFieldTypeZodSchema(config: ValidationSchemaPluginConfig, visitor: Visitor, field: InputValueDefinitionNode | FieldDefinitionNode, type: TypeNode, parentType?: TypeNode, isRoot = true, forceRequired = false, depthVariable?: string): string {
-  if (isListType(type)) {
-    const gen = generateFieldTypeZodSchema(config, visitor, field, type.type, type, false, false, depthVariable);
-    const arrayGen = `z.array(${maybeLazy(visitor, type.type, gen)})`;
-    const maybeDirectivesGen = isRoot ? applyDirectives(config, field, arrayGen) : arrayGen;
-    const maybeDefaultGen = hasNullDefault(field) ? maybeDirectivesGen : applyDefaultValue(config, visitor, field, type, maybeDirectivesGen);
-    if (!isNonNullType(parentType) && !forceRequired) {
-      if (hasNullDefault(field))
-        return withNullDefault(config, maybeDirectivesGen);
-
-      return `${maybeDefaultGen}.${zodOptionalType(config)}()`;
-    }
-    return maybeDefaultGen;
-  }
-  if (isNonNullType(type)) {
-    const gen = generateFieldTypeZodSchema(config, visitor, field, type.type, type, isRoot, forceRequired, depthVariable);
-    return maybeLazy(visitor, type.type, gen);
-  }
-  if (isNamedType(type)) {
-    const gen = generateNameNodeZodSchema(config, visitor, type.name, depthVariable);
-    if (isListType(parentType))
-      return `${gen}.nullable()`;
-
-    const appliedDirectivesGen = isRoot
-      ? hasNullDefault(field)
-        ? applyDirectives(config, field, gen)
-        : applyDefaultValue(config, visitor, field, type, applyDirectives(config, field, gen))
-      : gen;
-
-    if (isNonNullType(parentType)) {
-      if (visitor.shouldEmitAsNotAllowEmptyString(type.name.value))
-        return `${appliedDirectivesGen}.min(1)`;
-
-      return appliedDirectivesGen;
-    }
-    if (isListType(parentType))
-      return `${appliedDirectivesGen}.nullable()`;
-
-    if (forceRequired)
-      return appliedDirectivesGen;
-
-    return hasNullDefault(field)
-      ? withNullDefault(config, appliedDirectivesGen)
-      : `${appliedDirectivesGen}.${zodOptionalType(config)}()`;
-  }
-  console.warn('unhandled type:', type);
-  return '';
-}
-
-function isOneOfInputObject(node: InputObjectTypeDefinitionNode): boolean {
-  return node.directives?.some(directive => directive.name.value === 'oneOf') === true;
-}
-
 function hasDefaultValue(fields: readonly (FieldDefinitionNode | InputValueDefinitionNode)[]): boolean {
   return fields.some(field => field.kind === Kind.INPUT_VALUE_DEFINITION && field.defaultValue !== undefined);
-}
-
-function buildObjectExpression(config: ValidationSchemaPluginConfig, shape: string | undefined): string {
-  return ['z.object({', shape, `})${strictObjectSuffix(config)}`].join('\n');
-}
-
-function buildObjectReturn(config: ValidationSchemaPluginConfig, shape: string | undefined): string {
-  return [indent('return z.object({'), shape, indent(`})${strictObjectSuffix(config)}`)].join('\n');
-}
-
-function strictObjectSuffix(config: ValidationSchemaPluginConfig): string {
-  return config.strictObjectSchemas === true ? '.strict()' : '';
-}
-
-function zodOptionalType(config: ValidationSchemaPluginConfig): string {
-  return config.nullishBehavior ?? config.zodOptionalType ?? 'nullish';
-}
-
-function withNullDefault(config: ValidationSchemaPluginConfig, gen: string): string {
-  if (zodOptionalType(config) === 'optional')
-    return `${gen}.nullable().optional().default(null)`;
-
-  return `${gen}.${zodOptionalType(config)}().default(null)`;
-}
-
-function schemaDepthVariable(config: ValidationSchemaPluginConfig): string | undefined {
-  return typeof config.maxDepth === 'number' && config.validationSchemaExportType !== 'const'
-    ? 'depth'
-    : undefined;
-}
-
-function schemaDepthParameter(config: ValidationSchemaPluginConfig): string {
-  return schemaDepthVariable(config) ? 'depth = 0' : '';
-}
-
-function withDescription(config: ValidationSchemaPluginConfig, field: InputValueDefinitionNode | FieldDefinitionNode, gen: string): string {
-  if (config.withDescriptions !== true || !field.description?.value)
-    return gen;
-
-  return `${gen}.describe(${JSON.stringify(field.description.value)})`;
-}
-
-function applyDefaultValue(config: ValidationSchemaPluginConfig, visitor: Visitor, field: InputValueDefinitionNode | FieldDefinitionNode, type: TypeNode, gen: string): string {
-  if (field.kind !== Kind.INPUT_VALUE_DEFINITION || !field.defaultValue)
-    return gen;
-
-  return `${gen}.default(${defaultValueExpression(config, visitor, type, field.defaultValue)})`;
-}
-
-function defaultValueExpression(config: ValidationSchemaPluginConfig, visitor: Visitor, type: TypeNode, value: ConstValueNode): string {
-  if (value.kind === Kind.NULL)
-    return 'null';
-
-  if (isNonNullType(type))
-    return defaultValueExpression(config, visitor, type.type, value);
-
-  if (isListType(type)) {
-    if (value.kind === Kind.LIST)
-      return `[${value.values.map(item => defaultValueExpression(config, visitor, type.type, item)).join(', ')}]`;
-
-    return `[${defaultValueExpression(config, visitor, type.type, value)}]`;
-  }
-
-  if (isNamedType(type) && visitor.getType(type.name.value)?.astNode?.kind === 'EnumTypeDefinition' && value.kind === Kind.ENUM) {
-    if (!config.enumsAsTypes)
-      return `${enumDefaultTypeName(visitor, type)}.${enumDefaultValueName(config, value.value)}`;
-
-    return JSON.stringify(value.value);
-  }
-
-  if (isNamedType(type) && value.kind === Kind.OBJECT) {
-    const graphQLType = visitor.getType(type.name.value);
-    const astNode = graphQLType?.astNode;
-    if (astNode?.kind === Kind.INPUT_OBJECT_TYPE_DEFINITION && isInputObjectType(graphQLType)) {
-      const explicitFields = new Map(value.fields.map(field => [field.name.value, field.value]));
-      const fields = inputObjectFields(astNode, graphQLType.extensionASTNodes).flatMap((field) => {
-        const fieldValue = explicitFields.get(field.name.value) ?? field.defaultValue;
-        if (!fieldValue)
-          return [];
-
-        return `${field.name.value}: ${defaultValueExpression(config, visitor, field.type, fieldValue)}`;
-      });
-
-      return `{ ${fields.join(', ')} }`;
-    }
-  }
-
-  if (value.kind === Kind.INT || value.kind === Kind.FLOAT || value.kind === Kind.BOOLEAN)
-    return `${value.value}`;
-
-  if (value.kind === Kind.STRING)
-    return `"${escapeGraphQLCharacters(value.value)}"`;
-
-  return JSON.stringify(valueFromASTUntyped(value));
-}
-
-function hasNullDefault(field: InputValueDefinitionNode | FieldDefinitionNode): boolean {
-  return field.kind === Kind.INPUT_VALUE_DEFINITION && field.defaultValue?.kind === Kind.NULL;
-}
-
-function inputObjectFields(
-  astNode: InputObjectTypeDefinitionNode,
-  extensionASTNodes: readonly InputObjectTypeExtensionNode[] | undefined,
-): InputValueDefinitionNode[] {
-  return [
-    ...(astNode.fields ?? []),
-    ...(extensionASTNodes?.flatMap(extension => extension.fields ?? []) ?? []),
-  ];
-}
-
-function enumDefaultTypeName(visitor: Visitor, type: TypeNode): string {
-  if (isNonNullType(type))
-    return enumDefaultTypeName(visitor, type.type);
-
-  if (isNamedType(type))
-    return visitor.prefixTypeNamespace(visitor.convertSchemaName(type.name.value, visitor.getType(type.name.value)?.astNode?.kind));
-
-  return '';
-}
-
-function enumDefaultValueName(config: ValidationSchemaPluginConfig, value: string): string {
-  let enumValue = convertNameParts(value, resolveExternalModuleAndFn('change-case-all#pascalCase'), config.namingConvention?.transformUnderscore);
-
-  if (config.namingConvention?.enumValues)
-    enumValue = convertNameParts(value, resolveExternalModuleAndFn(config.namingConvention?.enumValues), config.namingConvention?.transformUnderscore);
-
-  return enumValue;
-}
-
-function applyDirectives(config: ValidationSchemaPluginConfig, field: InputValueDefinitionNode | FieldDefinitionNode, gen: string): string {
-  if (config.directives && field.directives) {
-    const formatted = formatDirectiveConfig(config.directives);
-    return gen + buildApi(formatted, field.directives);
-  }
-  return gen;
-}
-
-function generateNameNodeZodSchema(config: ValidationSchemaPluginConfig, visitor: Visitor, node: NameNode, depthVariable?: string): string {
-  const converter = visitor.getNameNodeConverter(node);
-
-  switch (converter?.targetKind) {
-    case 'InterfaceTypeDefinition':
-    case 'InputObjectTypeDefinition':
-    case 'ObjectTypeDefinition':
-    case 'UnionTypeDefinition':
-      // using switch-case rather than if-else to allow for future expansion
-      switch (config.validationSchemaExportType) {
-        case 'const':
-          return `${converter.convertName()}Schema`;
-        case 'function':
-        default:
-          if (
-            depthVariable
-            && (
-              converter.targetKind === 'InterfaceTypeDefinition'
-              || converter.targetKind === 'ObjectTypeDefinition'
-              || converter.targetKind === 'UnionTypeDefinition'
-            )
-          ) {
-            return `${depthVariable} >= ${config.maxDepth} ? ${anySchema} : ${converter.convertName()}Schema(${depthVariable} + 1)`;
-          }
-          return `${converter.convertName()}Schema()`;
-      }
-    case 'EnumTypeDefinition':
-      return `${converter.convertName()}Schema`;
-    case 'ScalarTypeDefinition':
-      return zod4Scalar(config, visitor, node.value);
-    default:
-      if (converter?.targetKind)
-        console.warn('Unknown targetKind', converter?.targetKind);
-
-      return zod4Scalar(config, visitor, node.value);
-  }
-}
-
-function maybeLazy(visitor: Visitor, type: TypeNode, schema: string): string {
-  if (!isNamedType(type)) {
-    return schema;
-  }
-
-  const schemaType = visitor.getType(type.name.value);
-  const isComplexType = !isScalarType(schemaType) && !isEnumType(schemaType);
-  return isComplexType ? `z.lazy(() => ${schema})` : schema;
-}
-
-function zod4Scalar(config: ValidationSchemaPluginConfig, visitor: Visitor, scalarName: string): string {
-  if (config.scalarSchemas?.[scalarName])
-    return config.scalarSchemas[scalarName];
-
-  const tsType = visitor.getScalarType(scalarName);
-  switch (tsType) {
-    case 'string':
-      return `z.string()`;
-    case 'number':
-      return `z.number()`;
-    case 'boolean':
-      return `z.boolean()`;
-  }
-
-  if (config.defaultScalarTypeSchema) {
-    return config.defaultScalarTypeSchema;
-  }
-
-  console.warn('unhandled scalar name:', scalarName);
-  return anySchema;
-}
-
-function unionLiterals(values: string[]): string {
-  if (values.length === 0)
-    return 'never';
-
-  return values.map(value => JSON.stringify(value)).join(' | ');
 }
